PycharmProjects/dv_compare_project/core_framework/requirements.txt
google-cloud-storage
google-cloud-pubsub
selenium
pandas
azure-devops
robotframework >= 3.1.1
robotframework-seleniumlibrary >= 3.3.1
openpyxl
pymsteams
protobuf >= 3.12.0